Move tests into ConnectionTraitTest
