+++ tests/Functional/AbstractFunctionalTestCase.php
@@ -9,7 +9,6 @@ use Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Driver\Mysqli\Driver as MysqliDriver;
 use Doctrine\DBAL\Driver\PDO\MySQL\Driver as PDODriver;
 use Doctrine\DBAL\DriverManager;
-use Doctrine\DBAL\Exception;
 use Facile\DoctrineMySQLComeBack\Tests\DeprecationTrait;
 use Facile\DoctrineMySQLComeBack\Tests\Functional\Spy\Connection;
 use Facile\DoctrineMySQLComeBack\Tests\Functional\Spy\PrimaryReadReplicaConnection;
@@ -17,7 +16,7 @@ use PHPUnit\Framework\TestCase;
 
 abstract class AbstractFunctionalTestCase extends TestCase
 {
-    private const UPDATE_QUERY = 'UPDATE test SET updatedAt = CURRENT_TIMESTAMP WHERE id = 1';
+    protected const UPDATE_QUERY = 'UPDATE test SET updatedAt = CURRENT_TIMESTAMP WHERE id = 1';
 
     use DeprecationTrait;
 
@@ -145,269 +144,6 @@ TABLE
             'PDO with savepoints' => [PDODriver::class, true],
             'PDO with no savepoints' => [PDODriver::class, false],
         ];
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testExecuteQueryShouldNotReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 0, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $this->expectException(Exception::class);
-
-        $connection->executeQuery('SELECT 1');
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testExecuteQueryShouldReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $connection->executeQuery('SELECT 1')->fetchAllNumeric();
-
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testQueryShouldReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $connection->executeQuery('SELECT 1');
-
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testExecuteUpdateShouldReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->createTestTable($connection);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $connection->executeStatement(self::UPDATE_QUERY);
-
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testExecuteStatementShouldReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->createTestTable($connection);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $connection->executeStatement(self::UPDATE_QUERY);
-
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldReconnectOnStatementExecuteError(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $statement = $connection->prepare("SELECT 'foo'");
-        $result = $statement->executeQuery()->fetchAllAssociative();
-
-        $this->assertSame([['foo' => 'foo']], $result);
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldResetStatementOnStatementExecuteError(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $statement = $connection->prepare("SELECT 'foo', ?, ?");
-        $statement->bindValue(1, 'bar');
-        $param = 'baz';
-        /** @psalm-suppress DeprecatedMethod */
-        $statement->bindParam(2, $param);
-        // change param by ref
-        $param = 'baz2';
-
-        $result = $statement->executeQuery()->fetchAllNumeric();
-
-        $this->assertSame([['foo', 'bar', $param]], $result);
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldReconnectOnStatementFetchAllAssociative(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $statement = $connection->prepare("SELECT 'foo'");
-        $result = $statement->executeQuery()->fetchAllAssociative();
-
-        $this->assertSame([['foo' => 'foo']], $result);
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldReconnectOnStatementFetchAllNumeric(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $statement = $connection->prepare("SELECT 'foo'");
-        $result = $statement->executeQuery()->fetchAllNumeric();
-
-        $this->assertSame([['0' => 'foo']], $result);
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testBeginTransactionShouldNotReconnectIfNested(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-
-        $connection->beginTransaction();
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $this->expectExceptionMessage('MySQL server has gone away');
-
-        $connection->beginTransaction();
-
-        if ($enableSavepoints) {
-            $this->fail('With savepoints enabled, test should fail without having to trigger a further query');
-        }
-
-        $connection->executeStatement('SELECT 1');
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testBeginTransactionShouldNotReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 0, $enableSavepoints);
-        $driver = $connection->getDriver();
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        if (is_a($driver, PDODriver::class)) {
-            $this->expectException(\PDOException::class);
-            $this->expectExceptionMessage('MySQL server has gone away');
-        }
-
-        $connection->beginTransaction();
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testBeginTransactionShouldReconnect(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $driver = $connection->getDriver();
-        $this->assertConnectionCount(1, $connection);
-        $this->forceDisconnect($connection);
-
-        $connection->beginTransaction();
-
-        if (is_a($driver, PDODriver::class)) {
-            $this->assertConnectionCount(2, $connection);
-        } else {
-            $this->assertConnectionCount(1, $connection);
-        }
-
-        $this->assertSame(1, $connection->getTransactionNestingLevel());
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldReconnectOnExecutePreparedStatement(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $statement = $connection->prepare('SELECT 1');
-
-        $this->forceDisconnect($connection);
-
-        $this->assertSame(1, $statement->executeStatement());
-        $this->assertConnectionCount(2, $connection);
-    }
-
-    /**
-     * @dataProvider driverDataProvider
-     *
-     * @param class-string<Driver> $driver
-     */
-    public function testShouldReconnectOnExecuteQueryPreparedStatement(string $driver, bool $enableSavepoints): void
-    {
-        $connection = $this->getConnectedConnection($driver, 1, $enableSavepoints);
-        $this->assertConnectionCount(1, $connection);
-        $statement = $connection->prepare('SELECT 1');
-
-        $this->forceDisconnect($connection);
-
-        $this->assertEquals([[1 => '1']], $statement->executeQuery()->fetchAllAssociative());
-        $this->assertConnectionCount(2, $connection);
     }
 
     /**
